Fix: Bug fixed

+++ models/event.go
@@ -7,12 +7,12 @@ import (
 )
 
 type Event struct {
-	ID          int
+	ID          int64
 	Name        string    `binding:"required"`
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserID      int
+	UserID      int64
 }
 
 func (e *Event) Save() error {
@@ -38,7 +38,7 @@ func (e *Event) Save() error {
 		return err
 	}
 
-	e.ID = int(eventId)
+	e.ID = int64(eventId)
 
 	return nil
 }
